Reworked entire project
Updated to use Sitch.co Boilerplate template.

Now using php and separating each style component in to a single .php
file in the _modules folder.

These files are then manually (for now) included in the index.php file
and the page is built.

Updated a lot of style guide specific styles.